Changes from refactor after hands on session
