+++ src/main/java/com/werfen/report/service/pdf/template/FormReportTemplateBuilder.java
@@ -7,6 +7,7 @@ import com.werfen.report.service.pdf.util.DesignTextBuilder;
 import net.sf.jasperreports.engine.design.JRDesignBand;
 import net.sf.jasperreports.engine.design.JRDesignSection;
 import net.sf.jasperreports.engine.design.JRDesignStaticText;
+import net.sf.jasperreports.engine.design.JasperDesign;
 import net.sf.jasperreports.engine.type.HorizontalTextAlignEnum;
 import net.sf.jasperreports.engine.type.ModeEnum;
 import net.sf.jasperreports.engine.type.SplitTypeEnum;
@@ -14,7 +15,7 @@ import net.sf.jasperreports.engine.type.VerticalTextAlignEnum;
 
 import java.awt.*;
 
-public class FormReportTemplateBuilder extends BaseReportTemplateBuilder {
+public class FormReportTemplateBuilder {
 
     public static final int SECTION_SEPARATOR_BEGIN = 5;
     public static final int SECTION_SEPARATOR_END = 10;
@@ -24,23 +25,24 @@ public class FormReportTemplateBuilder extends BaseReportTemplateBuilder {
     private static final String TITLE_FONT_NAME = "Helvetica-Bold";
     private static final int TITLE_HEIGHT = 23;
     private static final int TITLE_LEFT_PADDING = 9;
+    private static final int PAGE_MARGIN = 20;
 
-    public void addForm(FormReportData formReportData) {
-        formReportData.getSections().forEach(section -> this.addSection(section, 0));
+    public void addForm(JasperDesign jasperDesign, FormReportData formReportData) {
+        formReportData.getSections().forEach(section -> this.addSection(jasperDesign, section, 0));
     }
 
-    private void addSection(FormReportSection formReportSection, int sectionDepth) {
-        this.addTitle(formReportSection.getTitle(), sectionDepth);
+    private void addSection(JasperDesign jasperDesign, FormReportSection formReportSection, int sectionDepth) {
+        this.addTitle(jasperDesign, formReportSection.getTitle(), sectionDepth);
 
-        this.addSeparator(formReportSection.getFields().isEmpty() ? 0 : SECTION_SEPARATOR_BEGIN);
-        formReportSection.getFields().forEach(field -> this.addField(field, 0));
-        this.addSeparator(formReportSection.getFields().isEmpty() ? 0 : SECTION_SEPARATOR_END);
+        this.addSeparator(jasperDesign, formReportSection.getFields().isEmpty() ? 0 : SECTION_SEPARATOR_BEGIN);
+        formReportSection.getFields().forEach(field -> this.addField(jasperDesign, field, 0));
+        this.addSeparator(jasperDesign, formReportSection.getFields().isEmpty() ? 0 : SECTION_SEPARATOR_END);
 
-        formReportSection.getSubsections().forEach(subsection -> this.addSection(subsection, sectionDepth + 1));
+        formReportSection.getSubsections().forEach(subsection -> this.addSection(jasperDesign, subsection, sectionDepth + 1));
     }
 
-    private void addTitle(String title, int sectionDepth) {
-        final int BAND_WIDTH = this.jasperDesign.getPageWidth() - (PAGE_MARGIN * 2);
+    private void addTitle(JasperDesign jasperDesign, String title, int sectionDepth) {
+        final int BAND_WIDTH = jasperDesign.getPageWidth() - (PAGE_MARGIN * 2);
 
         JRDesignStaticText titleBox = new DesignTextBuilder()
                 .position(0, 0, BAND_WIDTH, TITLE_HEIGHT)
@@ -63,7 +65,7 @@ public class FormReportTemplateBuilder extends BaseReportTemplateBuilder {
         titleBand.addElement(titleBox);
         titleBand.addElement(titleText);
 
-        ((JRDesignSection) this.jasperDesign.getDetailSection()).addBand(titleBand);
+        ((JRDesignSection) jasperDesign.getDetailSection()).addBand(titleBand);
     }
 
     private Color getTitleColor(int sectionDepth) {
@@ -79,8 +81,8 @@ public class FormReportTemplateBuilder extends BaseReportTemplateBuilder {
         }
     }
 
-    private void addField(FormReportField formReportField, int fieldDepth) {
-        final int BAND_WIDTH = this.jasperDesign.getPageWidth() - (PAGE_MARGIN * 2);
+    private void addField(JasperDesign jasperDesign, FormReportField formReportField, int fieldDepth) {
+        final int BAND_WIDTH = jasperDesign.getPageWidth() - (PAGE_MARGIN * 2);
         JRDesignStaticText labelText = new DesignTextBuilder()
                 .position(FIELD_LABEL_X * (fieldDepth + 1), 0, BAND_WIDTH / 4, FIELD_HEIGHT)
                 .horizontalTextAlign(HorizontalTextAlignEnum.LEFT)
@@ -99,17 +101,17 @@ public class FormReportTemplateBuilder extends BaseReportTemplateBuilder {
         fieldBand.addElement(labelText);
         fieldBand.addElement(valueText);
 
-        ((JRDesignSection) this.jasperDesign.getDetailSection()).addBand(fieldBand);
+        ((JRDesignSection) jasperDesign.getDetailSection()).addBand(fieldBand);
 
-        formReportField.getSubfields().forEach(subfield -> this.addField(subfield, fieldDepth + 1));
+        formReportField.getSubfields().forEach(subfield -> this.addField(jasperDesign, subfield, fieldDepth + 1));
     }
 
-    private void addSeparator(int height) {
+    private void addSeparator(JasperDesign jasperDesign, int height) {
         if (height > 0) {
             JRDesignBand separatorBand = new JRDesignBand();
             separatorBand.setHeight(height);
             separatorBand.setSplitType(SplitTypeEnum.STRETCH);
-            ((JRDesignSection) this.jasperDesign.getDetailSection()).addBand(separatorBand);
+            ((JRDesignSection) jasperDesign.getDetailSection()).addBand(separatorBand);
         }
     }
 }
