+++ src/main/java/com/werfen/report/service/pdf/util/PdfExportService.java
@@ -1,0 +1,40 @@
+package com.werfen.report.service.pdf.util;
+
+import net.sf.jasperreports.engine.JRException;
+import net.sf.jasperreports.engine.JasperPrint;
+import net.sf.jasperreports.engine.export.JRPdfExporter;
+import net.sf.jasperreports.export.SimpleExporterInput;
+import net.sf.jasperreports.export.SimpleOutputStreamExporterOutput;
+import net.sf.jasperreports.export.SimplePdfExporterConfiguration;
+import net.sf.jasperreports.export.SimplePdfReportConfiguration;
+
+import java.io.File;
+
+public class PdfExportService {
+
+    public File export(JasperPrint jasperPrint, String filePath) throws JRException {
+        JRPdfExporter exporter = new JRPdfExporter();
+        exporter.setExporterInput(new SimpleExporterInput(jasperPrint));
+        exporter.setExporterOutput(new SimpleOutputStreamExporterOutput(filePath));
+        exporter.setConfiguration(getPdfReportConfiguration());
+        exporter.setConfiguration(getPdfExporterConfiguration());
+
+        exporter.exportReport();
+        return new File(filePath);
+    }
+
+    private SimplePdfReportConfiguration getPdfReportConfiguration() {
+        SimplePdfReportConfiguration reportConfig = new SimplePdfReportConfiguration();
+        reportConfig.setSizePageToContent(true);
+        reportConfig.setForceLineBreakPolicy(false);
+        return reportConfig;
+    }
+
+    private SimplePdfExporterConfiguration getPdfExporterConfiguration() {
+        SimplePdfExporterConfiguration exportConfig = new SimplePdfExporterConfiguration();
+        exportConfig.setEncrypted(true);
+        exportConfig.setAllowedPermissionsHint("PRINTING");
+        return exportConfig;
+    }
+
+}
